+++ tests/engine/cloud/test_utilities.py
@@ -30,6 +30,19 @@ def test_preparing_state_for_cloud_replaces_cached_inputs_with_safe(cls):
 
 
 @pytest.mark.parametrize("cls", [s for s in all_states if s.__name__ != "State"])
+def test_preparing_state_for_cloud_ignores_the_lack_of_result_handlers_for_cached_inputs(
+    cls,
+):
+    xres = Result(3, result_handler=None)
+    state = prepare_state_for_cloud(cls(cached_inputs=dict(x=xres)))
+    assert isinstance(state, cls)
+    assert state.result is None
+    assert state._result == NoResult
+    assert state.cached_inputs == dict(x=xres)
+    assert state.serialize()["cached_inputs"]["x"]["type"] == "NoResultType"
+
+
+@pytest.mark.parametrize("cls", [s for s in all_states if s.__name__ != "State"])
 def test_preparing_state_for_cloud_does_nothing_if_result_is_none(cls):
     xres = Result(None, result_handler=JSONResultHandler())
     state = prepare_state_for_cloud(cls(cached_inputs=dict(x=xres)))
@@ -40,6 +53,7 @@ def test_preparing_state_for_cloud_does_nothing_if_result_is_none(cls):
     assert state.serialize()["cached_inputs"]["x"]["type"] == "NoResultType"
 
 
+@pytest.mark.skip(reason="Result Handlers are not required to exist currently")
 @pytest.mark.parametrize("cls", [s for s in all_states if s.__name__ != "State"])
 def test_preparing_state_for_cloud_fails_if_cached_inputs_have_no_handler(cls):
     xres = Result(3, result_handler=None)